Render two fields as the live API actually sends them
`priceWithSellerFee` was `int` while prod sends 25.22 — and the spec itself is
self-contradictory here, declaring the field `integer` in some places and
`number` in others. `restore_percents` was `str` while prod sends 0; the spec
says `integer` and is right, so that one was a rendering bug, not a spec bug.

Both were already hand-patched on individual seller models back in July, and
the purchase path was missed — which is the argument for fixing it in the
generator instead: a hand-patch fixes the one file someone happened to hit, and
`restore_percents` is rendered into several models. Doing it here also keeps
those files generated; a hand-patch means dropping the auto-gen marker, which
freezes the file against every later spec change.

Once the three seller models agreed, the dedup pass collapsed them into one —
hence the deletion.

This does NOT yet make `purchasing_check` parse. Measured against prod the same
day, the response also omits ~11 fields the model marks required (ask_user_id,
ask_date, discount_price, min_price, deposit, bumpSettings, ...), all failing
as `missing` rather than as a type error. That is the known "spec said
required, live response omits it" class and wants a decision about response
DTOs in general, not another per-field entry.

+++ src/pylzt/models/market/mihoyo_item.py
@@ -68,7 +68,7 @@ class MihoyoItem(BaseItem):
     mihoyo_zenless_currency: int
     feedback_data: str
     isIgnored: bool
-    priceWithSellerFee: int
+    priceWithSellerFee: float
     guarantee: str
     canViewLoginData: bool
     canUpdateItemStats: bool
